+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [master](https://github.com/arangodb/kube-arangodb/tree/master) (N/A)
 - (Maintenance) Bump Helm to v3.21.4, dropping the containerd and golang.org/x/crypto/openpgp dependencies (clears GO-2026-5932/5622/5338/5064)
+- (Feature) (Platform) Add ArangoPlatformWorkflow CRD
 - (Maintenance) Bump Go toolchain to 1.26.8 and fix go vet printf findings in logging calls surfaced by it
 - (Maintenance) Bump Kubernetes libraries to 1.34.11
 - (Maintenance) Bump Kubernetes libraries to 1.35.8